requirements.txt
albumentations
lightning
matplotlib
monai
numpy
nvitop
opencv-python
open-clip-torch
pandas
pillow
PyYAML
scikit-learn
seaborn
SimpleITK
tabulate
tensorboard
timm
torch
torch_flops
torchmetrics
torchvision
tqdm
wandb
ipykernel